Remove duplicate single token mentions
Since we operate on subword-tokens we can have two, in the models
view, distinct mentions cover the exact same conll tokens.
The official conll scorer doesn't like duplicates and exits early if too
many are present.
For those mentions that only cover a single conll token we can easily
remove duplicates. This just means that, for example, "Rheinland-Pfalz"
is not present as multiple distinct mentions for its three sub-tokens.