Allow smaller embedding layer in weight than config specifies
For the distance related embedding layers we now randomly initialize
those values that are required by the config but not found in the
weights.

This does have the advantage of being able to load old models while
using a larger size for the embedding. Specifically this is useful when
loading weights from training on news texts.